✨ [feat] コマンド実行時のオプションを追加し、許可されたコマンドと作業ディレクトリを設定可能に
